Add validation for phone, styling for email
Add in a placeholder to indicate the format of the phone number to the
user. Then add in a pattern for regular expression to match this format.
Add in validation CSS to indicate when it is not invalid. Do the same
for email with different wording on the message since email already
has a built in validation